Internal reorg and docstring cleanup.
