+++ filters/test.py
@@ -11,14 +11,19 @@ from unittest import TestCase
 from collections import OrderedDict
 from six import iterkeys, string_types
 
-from filters import BaseFilter, FilterRunner
+from filters.base import BaseFilter
+from filters.handlers import FilterRunner
+
+__all__ = [
+    'BaseFilterTestCase',
+]
 
 
 def sorted_dict(value):
     # type: (Mapping) -> Any
     """
     Sorts a dict's keys to avoid leaking information about the
-        backend's handling of unordered dicts.
+    backend's handling of unordered dicts.
     """
     if isinstance(value, Mapping):
         return OrderedDict(
@@ -38,43 +43,44 @@ class BaseFilterTestCase(TestCase):
     Base functionality for request filter unit tests.
 
     Prevents typos from causing invalid test passes/failures by
-        abstracting the Filter type out of filtering ops; just set
-        `filter_type` at the top of your test case, and then use
-        `assertFilterPasses` and `assertFilterErrors` to check use
-        cases.
+    abstracting the Filter type out of filtering ops; just set
+    ``filter_type`` at the top of your test case, and then use
+    ``assertFilterPasses`` and ``assertFilterErrors`` to check use
+    cases.
     """
     filter_type = None # type: Callable[..., BaseFilter]
 
     class unmodified(object):
         """
-        Used by `assertFilterPasses` so that you can omit the
-            `expected_value` parameter.
+        Used by ``assertFilterPasses`` so that you can omit the
+        ``expected_value`` parameter.
         """
         pass
 
     class skip_value_check(object):
         """
-        Used by `assertFilterPasses` to skip checking the filtered
-            value.  This is useful for tests where a simple equality
-            check is not practical.
+        Used by ``assertFilterPasses`` to skip checking the filtered
+        value.  This is useful for tests where a simple equality
+        check is not practical.
 
-        Note:  If you use `skip_value_check`, you should add extra
-            assertions to your test to make sure the filtered value
-            conforms to expectations.
+        Note:  If you use ``skip_value_check``, you should add extra
+        assertions to your test to make sure the filtered value
+        conforms to expectations.
         """
         pass
 
     def assertFilterPasses(self, runner, expected_value=unmodified):
         """
         Asserts that the FilterRunner returns the specified value,
-            without errors.
+        without errors.
 
-        :param runner: Usually a FilterRunner instance, but you can
-            pass in the test value itself if you want (it will
-            automatically be run through `_filter`).
+        :param runner:
+            Usually a FilterRunner instance, but you can pass in the
+            test value itself if you want (it will automatically be run
+            through ``_filter``).
 
-        :param expected_value: The expected value for
-            `runner.cleaned_data`.
+        :param expected_value:
+            The expected value for ``runner.cleaned_data``.
 
             If omitted, the assertion will check that the incoming
             value is returned unmodified.
@@ -84,16 +90,16 @@ class BaseFilterTestCase(TestCase):
     def assertFilterErrors(self, runner, expected_codes, expected_value=None):
         """
         Asserts that the FilterRunner generates the specified error
-            codes.
+        codes.
 
-        :param runner: Usually a FilterRunner instance, but you can
-            pass in the test value itself if you want (it will
-            automatically be run through `_filter`).
+        :param runner:
+            Usually a FilterRunner instance, but you can pass in the
+            test value itself if you want (it will automatically be
+            run through `_filter`).
 
-        :type expected_codes: dict[unicode, list[unicode]]|list[unicode]
-
-        :param expected_value: Expected value for the FilterRunner's
-            `cleaned_data` property (usually `None`).
+        :param expected_value:
+            Expected value for ``runner.cleaned_data`` (usually
+            ``None``).
         """
         if not isinstance(runner, FilterRunner):
             runner = self._filter(runner) # type: FilterRunner
@@ -145,9 +151,9 @@ class BaseFilterTestCase(TestCase):
         Applies the Filter to the specified value.
 
         Generally, you don't have to use this method in your test case,
-            unless you want to specify Filter options.
+        unless you want to specify Filter options.
 
-        Example:
+        Example::
 
             self.filter_type = Min
 
@@ -157,10 +163,11 @@ class BaseFilterTestCase(TestCase):
             # Min(min_val=40).apply(42)
             self.assertFilterPasses(self._filter(42, min_val=40))
 
-        :param args:    Positional params to pass to the Filter's
-            initializer.
-        :param kwargs:  Keyword params to pass to the Filter's
-            initializer.
+        :param args:
+            Positional params to pass to the Filter's initializer.
+
+        :param kwargs:
+            Keyword params to pass to the Filter's initializer.
         """
         assert len(args) > 0, 'First argument must be the filtered value.'
 
@@ -173,10 +180,10 @@ class BaseFilterTestCase(TestCase):
     def _check_filter_value(self, cleaned_data, expected):
         """
         Checks the value returned by the Filter, used by
-            `assertFilterPasses`.
+        ``assertFilterPasses``.
 
         In certain cases, it may be useful to override this method in
-            your test case subclass.
+        your test case subclass.
 
         :param cleaned_data: FilterRunner.cleaned_data
         """
