+++ test/simple_test.py
@@ -2,11 +2,9 @@
 from __future__ import absolute_import, division, print_function, \
     unicode_literals
 
-from collections import Sequence
 from datetime import datetime, date
-from decimal import Decimal
-from xml.etree.ElementTree import Element
 
+from collections import Sequence
 from dateutil.tz import tzoffset
 from pytz import utc
 from six import PY2, binary_type, python_2_unicode_compatible, text_type
@@ -276,200 +274,6 @@ class ByteArrayTestCase(BaseFilterTestCase):
         self.assertFilterErrors(
             self._filter(value, encoding='latin-1'),
             [f.ByteArray.CODE_BAD_ENCODING],
-        )
-
-
-class ByteStringTestCase(BaseFilterTestCase):
-    filter_type = f.ByteString
-
-    def test_pass_none(self):
-        """
-        `None` always passes this Filter.
-
-        Use `Required | ByteString` if you want to reject `None`.
-        """
-        self.assertFilterPasses(None)
-
-    def test_pass_unicode(self):
-        """
-        The incoming value is a unicode.
-        """
-        self.assertFilterPasses(
-            'Iñtërnâtiônàlizætiøn',
-
-            # 'Iñtërnâtiônàlizætiøn' encoded as bytes using utf-8:
-            b'I\xc3\xb1t\xc3\xabrn\xc3\xa2ti\xc3'
-            b'\xb4n\xc3\xa0liz\xc3\xa6ti\xc3\xb8n',
-        )
-
-    def test_pass_bytes_utf8(self):
-        """
-        The incoming value is a byte string, already encoded as UTF-8.
-        """
-        self.assertFilterPasses(
-            b'I\xc3\xb1t\xc3\xabrn\xc3\xa2ti\xc3'
-            b'\xb4n\xc3\xa0liz\xc3\xa6ti\xc3\xb8n'
-        )
-
-    def test_fail_bytes_non_utf8(self):
-        """
-        The incoming value is a byte string, but encoded using a
-            different codec.
-        """
-        # 'Iñtërnâtiônàlizætiøn' encoded as bytes using ISO-8859-1:
-        incoming = b'I\xf1t\xebrn\xe2ti\xf4n\xe0liz\xe6ti\xf8n'
-
-        self.assertFilterErrors(
-            incoming,
-            [f.ByteString.CODE_DECODE_ERROR],
-        )
-
-        # In order for this to work, we have to tell the Filter what
-        #   encoding to use:
-        self.assertFilterPasses(
-            self._filter(incoming, encoding='iso-8859-1'),
-
-            # The result is re-encoded using UTF-8.
-            b'I\xc3\xb1t\xc3\xabrn\xc3\xa2ti\xc3'
-            b'\xb4n\xc3\xa0liz\xc3\xa6ti\xc3\xb8n',
-        )
-
-    def test_pass_string_like_object(self):
-        """
-        The incoming value is an object that can be cast as a unicode.
-        """
-        self.assertFilterPasses(
-            Unicody('༼ つ ◕_◕ ༽つ'),
-
-            # Stoned Kirby?  Jigglypuff in dance mode?
-            # I have no idea what this is.
-            b'\xe0\xbc\xbc \xe3\x81\xa4 \xe2\x97\x95_'
-            b'\xe2\x97\x95 \xe0\xbc\xbd\xe3\x81\xa4',
-        )
-
-    def test_pass_bytes_like_object(self):
-        """
-        The incoming value is an object that can be cast as a byte
-            string.
-        """
-        value = (
-            # Person
-            b'(\xe2\x95\xaf\xc2\xb0\xe2\x96\xa1\xc2\xb0)'
-            # Particle Effects
-            b'\xe2\x95\xaf\xef\xb8\xb5 '
-            # Table
-            b'\xe2\x94\xbb\xe2\x94\x81\xe2\x94\xbb'
-        )
-
-        self.assertFilterPasses(Bytesy(value), value)
-
-    def test_pass_boolean(self):
-        """
-        The incoming value is a boolean (treated as an int).
-        """
-        self.assertFilterPasses(True, b'1')
-
-    def test_pass_decimal_with_scientific_notation(self):
-        """
-        The incoming value is a Decimal that was parsed from scientific
-            notation.
-        """
-        # Note that `str(Decimal('2.8E6')` yields b'2.8E+6', which is
-        #   not what we want!
-        self.assertFilterPasses(
-            Decimal('2.8E6'),
-            b'2800000',
-        )
-
-    def test_pass_xml_element(self):
-        """The incoming value is an ElementTree XML Element."""
-        self.assertFilterPasses(
-            Element('foobar'),
-            b'<foobar />',
-        )
-
-    def test_unicode_normalization_off_by_default(self):
-        """
-        By default, the Filter does not apply normalization before
-            encoding.
-
-        :see: https://en.wikipedia.org/wiki/Unicode_equivalence
-        :see: http://stackoverflow.com/q/16467479
-        """
-        self.assertFilterPasses(
-            #   U+0065 LATIN SMALL LETTER E
-            # + U+0301 COMBINING ACUTE ACCENT
-            # (2 code points)
-            'Ame\u0301lie',
-
-            # Result is the same string, encoded using UTF-8.
-            b'Ame\xcc\x81lie',
-        )
-
-    def test_unicode_normalization_forced(self):
-        """
-        You can force the Filter to apply normalization before encoding.
-
-        :see: https://en.wikipedia.org/wiki/Unicode_equivalence
-        :see: http://stackoverflow.com/q/16467479
-        """
-        self.assertFilterPasses(
-            self._filter(
-                # Same decomposed sequence from previous test...
-                'Ame\u0301lie',
-
-                # ... but this time we tell the Filter to normalize the
-                #   value before encoding it.
-                normalize = True,
-            ),
-
-            # U+00E9 LATIN SMALL LETTER E WITH ACUTE
-            # (1 code point, encoded as bytes)
-            b'Am\xc3\xa9lie',
-        )
-
-    def test_remove_non_printables_off_by_default(self):
-        """
-        By default, the Filter does not remove non-printable
-            characters.
-        """
-        self.assertFilterPasses(
-            # \u0000-\u001f are ASCII control characters.
-            # \uffff is a Unicode control character.
-            '\u0010Hell\u0000o,\u001f wor\uffffld!',
-
-            b'\x10Hell\x00o,\x1f wor\xef\xbf\xbfld!',
-        )
-
-    def test_remove_non_printables_forced(self):
-        """
-        You can force the Filter to remove non-printable characters
-            before encoding.
-        """
-        self.assertFilterPasses(
-            self._filter(
-                '\u0010Hell\u0000o,\u001f wor\uffffld!',
-                normalize = True,
-            ),
-            b'Hello, world!',
-        )
-
-    def test_newline_normalization_off_by_default(self):
-        """
-        By default, the Filter does not normalize line endings.
-        """
-        self.assertFilterPasses(
-            'unix\n - windows\r\n - weird\r\r\n',
-            b'unix\n - windows\r\n - weird\r\r\n',
-        )
-
-    def test_newline_normalization_forced(self):
-        """
-        You can force the Filter to normalize line endings.
-        """
-        self.assertFilterPasses(
-            self._filter('unix\n - windows\r\n - weird\r\r\n', normalize=True),
-            b'unix\n - windows\n - weird\n\n',
         )
 
 
@@ -800,121 +604,6 @@ class DatetimeTestCase(BaseFilterTestCase):
         )
 
 
-class DecimalTestCase(BaseFilterTestCase):
-    filter_type = f.Decimal
-
-    def test_pass_none(self):
-        """
-        `None` always passes this Filter.
-
-        Use `Required | Decimal` if you want to reject `None`.
-        """
-        self.assertFilterPasses(None)
-
-    def test_pass_valid_decimal(self):
-        """The incoming value can be interpreted as a Decimal."""
-        value = '3.1415926'
-
-        self.assertFilterPasses(value, Decimal(value))
-
-    def test_pass_max_precision(self):
-        """
-        You can limit the max precision of the resulting Decimal
-            object.
-
-        Note that a too-precise value is not considered invalid;
-            instead, it gets rounded to the expected precision.
-        """
-        self.assertFilterPasses(
-            self._filter('3.1415926', max_precision=3),
-            Decimal('3.142'),
-        )
-
-    def test_pass_zero(self):
-        """
-        0 is also considered a valid Decimal value.
-        """
-        value = '0'
-        self.assertFilterPasses(value, Decimal(value))
-
-    def test_pass_scientific_notation(self):
-        """
-        Scientific notation is considered valid, as in certain cases it
-            may be the only way to represent a really large or small
-            value.
-        """
-        value = '2.8E-12'
-        self.assertFilterPasses(value, Decimal(value))
-
-    def test_pass_boolean(self):
-        """
-        Booleans are technically ints, so they can be converted to
-            Decimals.
-        """
-        value = True
-        self.assertFilterPasses(value, Decimal(value))
-
-    def test_fail_invalid_value(self):
-        """
-        The incoming value cannot be converted to a Decimal.
-        """
-        self.assertFilterErrors(
-            'this is not a decimal',
-            [f.Decimal.CODE_INVALID],
-        )
-
-    def test_fail_non_finite_value(self):
-        """
-        Non-finite values like 'NaN' and 'Inf' are considered invalid,
-            even though they are technically parseable.
-        """
-        self.assertFilterErrors( 'NaN', [f.Decimal.CODE_NON_FINITE])
-        self.assertFilterErrors('+Inf', [f.Decimal.CODE_NON_FINITE])
-        self.assertFilterErrors('-Inf', [f.Decimal.CODE_NON_FINITE])
-        # There are a few other possible non-finite values out there,
-        #   but you get the idea.
-
-    def test_pass_tuple(self):
-        """
-        You may pass a 3-tuple for more control over how the resulting
-            Decimal object is created.
-        """
-        value = (0, (4, 2), -1)
-        self.assertFilterPasses(value, Decimal(value))
-
-    def test_fail_tuple_invalid(self):
-        """
-        If you're going to use a tuple, you've got to make sure you get
-            it right!
-        """
-        self.assertFilterErrors(('1', '2', '3'), [f.Decimal.CODE_INVALID])
-
-    def test_fail_tuple_disallowed(self):
-        """
-        The filter is configured to disallow tuple values.
-        """
-        self.assertFilterErrors(
-            self._filter((0, (4, 2), -1), allow_tuples=False),
-            [f.Type.CODE_WRONG_TYPE],
-        )
-
-    def test_fail_bytes(self):
-        """
-        To ensure that the filter behaves the same in Python 2 and
-          Python 3, bytes are not allowed.
-        """
-        self.assertFilterErrors(b'-12', [f.Type.CODE_WRONG_TYPE])
-
-    def test_fail_unsupported_type(self):
-        """
-        The incoming value has an unsupported type.
-        """
-        self.assertFilterErrors(
-            {0, (4, 2), -1},
-            [f.Type.CODE_WRONG_TYPE],
-        )
-
-
 class EmptyTestCase(BaseFilterTestCase):
     filter_type = f.Empty
 
@@ -971,84 +660,6 @@ class EmptyTestCase(BaseFilterTestCase):
             represents SOME kind of value.
         """
         self.assertFilterErrors(False, [f.Empty.CODE_NOT_EMPTY])
-
-
-class IntTestCase(BaseFilterTestCase):
-    filter_type = f.Int
-
-    def test_pass_none(self):
-        """
-        `None` always passes this Filter.
-
-        Use `Required | Int` if you want to reject `None`.
-        """
-        self.assertFilterPasses(None)
-
-    def test_pass_valid_int(self):
-        """The incoming value can be interpreted as an int."""
-        self.assertFilterPasses('42', 42)
-
-    def test_pass_zero(self):
-        """The incoming value is zero."""
-        self.assertFilterPasses('0', 0)
-
-    def test_pass_negative(self):
-        """The incoming value is a negative int."""
-        self.assertFilterPasses('-314', -314)
-
-    def test_pass_boolean(self):
-        """Booleans are technically ints."""
-        self.assertFilterPasses(True, 1)
-
-    def test_fail_invalid_value(self):
-        """
-        The incoming value cannot be interpreted as a number, let alone
-            an integer.
-        """
-        self.assertFilterErrors(
-            'this is not an int',
-            [f.Decimal.CODE_INVALID],
-        )
-
-    def test_fail_bytes(self):
-        """
-        To ensure that the filter behaves the same in Python 2 and
-          Python 3, bytes are not allowed.
-        """
-        self.assertFilterErrors(b'-12', [f.Type.CODE_WRONG_TYPE])
-
-    def test_fail_float_value(self):
-        """
-        The incoming value contains significant digits after the
-            decimal point.
-        """
-        self.assertFilterErrors(
-            '42.01',
-            [f.Int.CODE_DECIMAL],
-        )
-
-    def test_pass_int_point_zero(self):
-        """
-        The incoming value contains only insignificant digits after the
-            decimal point.
-        """
-        self.assertFilterPasses('42.0000000000000', 42)
-
-    def test_pass_scientific_notation(self):
-        """The incoming value is expressed in scientific notation."""
-        self.assertFilterPasses('2.6E4', 26000)
-
-    def test_fail_non_finite_value(self):
-        """The incoming value is a non-finite value."""
-        self.assertFilterErrors( 'NaN', [f.Decimal.CODE_NON_FINITE])
-        self.assertFilterErrors('+Inf', [f.Decimal.CODE_NON_FINITE])
-        self.assertFilterErrors('-Inf', [f.Decimal.CODE_NON_FINITE])
-        # There are a few other possible non-finite values out there,
-        #   but you get the idea.
-
-    def test_pass_int(self):
-        """The incoming value is already an int object."""
-        self.assertFilterPasses(777)
 
 
 class MaxLengthTestCase(BaseFilterTestCase):
@@ -1495,185 +1106,4 @@ class TypeTestCase(BaseFilterTestCase):
         self.assertFilterErrors(
             self._filter(text_type, allowed_types=text_type),
             [f.Type.CODE_WRONG_TYPE],
-        )
-
-
-class UnicodeTestCase(BaseFilterTestCase):
-    filter_type = f.Unicode
-
-    def test_pass_none(self):
-        """
-        `None` always passes this Filter.
-
-        Use `Required | Unicode` if you want to reject `None`.
-        """
-        self.assertFilterPasses(None)
-
-    def test_pass_unicode(self):
-        """The incoming value is a unicode."""
-        self.assertFilterPasses('┻━┻︵ \(°□°)/ ︵ ┻━┻ ') # RAWR!
-
-    def test_pass_bytes_utf8(self):
-        """
-        The incoming value is a byte string that is encoded as UTF-8.
-        """
-        self.assertFilterPasses(
-            # You get used to it.
-            # I don't even see the code.
-            # All I see is, "blond"... "brunette"... "redhead"...
-            # Hey, you uh... want a drink?
-            b'\xe2\x99\xaa '
-            b'\xe2\x94\x8f(\xc2\xb0.\xc2\xb0)\xe2\x94\x9b '
-            b'\xe2\x94\x97(\xc2\xb0.\xc2\xb0)\xe2\x94\x93 '
-            b'\xe2\x99\xaa',
-
-            '♪ ┏(°.°)┛ ┗(°.°)┓ ♪',
-        )
-
-    def test_fail_bytes_non_utf8(self):
-        """
-        The incoming value is a byte string that is encoded using a
-            codec other than UTF-8.
-
-        Note that there is no such thing as a unicode object with
-            the "wrong encoding".
-
-        :see: https://docs.python.org/2/howto/unicode.html
-        """
-        # How about something a bit more realistic for this test?
-        #   Like the Swedish word for 'Apple'.
-        # noinspection SpellCheckingInspection
-        incoming = b'\xc4pple'
-
-        self.assertFilterErrors(
-            incoming,
-            [f.Unicode.CODE_DECODE_ERROR],
-        )
-
-        # In order for this to work, we have to tell the Filter what
-        #   encoding to use:
-        self.assertFilterPasses(
-            self._filter(incoming, encoding='iso-8859-1'),
-            'Äpple',
-        )
-
-    def test_pass_string_like_object(self):
-        """
-        The incoming value is an object that can be cast as a string.
-        """
-        value = '／人 ⌒ ‿‿ ⌒ 人＼' # Squee!
-
-        self.assertFilterPasses(
-            Unicody(value),
-            value,
-        )
-
-    def test_pass_bytes_like_object(self):
-        """
-        The incoming value is an object that can be cast as a byte
-            string.
-        """
-        self.assertFilterPasses(
-            Bytesy(b'(\xe2\x99\xa5\xe2\x80\xbf\xe2\x99\xa5)'),
-
-            # I can almost hear the sappy music now.
-            '(♥‿♥)',
-        )
-
-    def test_pass_boolean(self):
-        """The incoming value is a boolean (treated as an int)."""
-        self.assertFilterPasses(True, '1')
-
-    def test_pass_decimal_with_scientific_notation(self):
-        """
-        The incoming value is a Decimal that was parsed from scientific
-            notation.
-        """
-        # Note that `text_type(Decimal('2.8E6')` yields '2.8E+6', which
-        #   is not what we want!
-        self.assertFilterPasses(
-            Decimal('2.8E6'),
-            '2800000',
-        )
-
-    def test_pass_xml_element(self):
-        """The incoming value is an ElementTree XML Element."""
-        self.assertFilterPasses(
-            Element('foobar'),
-            '<foobar />',
-        )
-
-    def test_unicode_normalization(self):
-        """
-        The Filter always returns the NFC form of the unicode string.
-
-        :see: https://en.wikipedia.org/wiki/Unicode_equivalence
-        :see: http://stackoverflow.com/q/16467479
-        """
-        #   U+0065 LATIN SMALL LETTER E
-        # + U+0301 COMBINING ACUTE ACCENT
-        # (2 code points)
-        decomposed  = 'Ame\u0301lie'
-
-        # U+00E9 LATIN SMALL LETTER E WITH ACUTE
-        # (1 code point)
-        composed    = 'Am\xe9lie'
-
-        self.assertFilterPasses(decomposed, composed)
-
-    def test_unicode_normalization_disabled(self):
-        """You can force the Filter not to perform normalization."""
-        decomposed  = 'Ame\u0301lie'
-
-        self.assertFilterPasses(
-            self._filter(decomposed, normalize=False),
-            decomposed,
-        )
-
-    def test_remove_non_printables(self):
-        """
-        By default, this Filter also removes non-printable characters
-            (both ASCII and Unicode varieties).
-        """
-        self.assertFilterPasses(
-            # \x00-\x1f are ASCII control characters.
-            # \xef\xbf\xbf is the Unicode control character \uffff,
-            #   encoded as UTF-8.
-            b'\x10Hell\x00o,\x1f wor\xef\xbf\xbfld!',
-
-            'Hello, world!',
-        )
-
-    def test_remove_non_printables_disabled(self):
-        """
-        You can force the Filter not to remove non-printable characters.
-        """
-        self.assertFilterPasses(
-            self._filter(
-                b'\x10Hell\x00o,\x1f wor\xef\xbf\xbfld!',
-                normalize = False,
-            ),
-
-            '\u0010Hell\u0000o,\u001f wor\uffffld!',
-        )
-
-    def test_newline_normalization(self):
-        """
-        By default, any newlines in the string are automatically
-            converted to unix-style ('\n').
-        """
-        self.assertFilterPasses(
-            b'unix\n - windows\r\n - weird\r\r\n',
-            'unix\n - windows\n - weird\n\n',
-        )
-
-    def test_newline_normalization_disabled(self):
-        """You can force the Filter not to normalize line endings."""
-        self.assertFilterPasses(
-            self._filter(
-                b'unix\n - windows\r\n - weird\r\r\n',
-                normalize = False,
-            ),
-
-            'unix\n - windows\r\n - weird\r\r\n',
         )
